Refactored file and package structure
