+++ src/main/java/com/softserve/skillscope/config/authentication/AuthenticationService.java
@@ -1,7 +1,7 @@
 package com.softserve.skillscope.config.authentication;
 
-import com.softserve.skillscope.talent.model.dto.responce.JwtToken;
-import com.softserve.skillscope.talent.model.dto.request.RegistrationRequest;
+import com.softserve.skillscope.talent.model.responce.JwtToken;
+import com.softserve.skillscope.talent.model.request.RegistrationRequest;
 
 public interface AuthenticationService {
     JwtToken registration(RegistrationRequest registrationRequest);
