+++ src/main/java/com/softserve/skillscope/config/authentication/AuthenticationServiceImpl.java
@@ -3,8 +3,8 @@ package com.softserve.skillscope.config.authentication;
 import com.softserve.skillscope.exception.talentException.TalentAlreadyExistsException;
 import com.softserve.skillscope.exception.talentException.TalentNotFoundException;
 import com.softserve.skillscope.talent.TalentRepository;
-import com.softserve.skillscope.talent.model.dto.responce.JwtToken;
-import com.softserve.skillscope.talent.model.dto.request.RegistrationRequest;
+import com.softserve.skillscope.talent.model.responce.JwtToken;
+import com.softserve.skillscope.talent.model.request.RegistrationRequest;
 import com.softserve.skillscope.talent.model.entity.Talent;
 import com.softserve.skillscope.talentInfo.model.entity.TalentInfo;
 import jakarta.transaction.Transactional;
